Added vector normalization to sentence embeddings
