+++ .flake8
@@ -1,3 +1,3 @@
 [flake8]
-extend-ignore = E226
+extend-ignore = E226, F405, F403, W605
 max-line-length=160
