+++ text2emoji/data/random_embeddings.py
@@ -1,0 +1,32 @@
+import numpy as np
+import pandas as pd
+
+
+def generate_embeddings(filename, n_dimensions, size):
+    """
+    Generate random embeddings and labels and save them as csv files.
+    Only used for testing purposes. The real embeddings are generated by a
+    separate model
+
+    Args:
+        filename (string): name of the file to save data to
+        n_dimensions (int): number of dimensions of the embeddings
+        size (int): number of samples to generate
+    """
+
+    # Generate size random embeddings of n_dimensions dimensions
+    embeddings = np.random.rand(size, n_dimensions)
+
+    # Generate random labels from 0 to 19
+    labels = np.random.randint(0, 20, size=size)
+
+    # Save as csv, combine embeddings and labels
+    df = pd.DataFrame(embeddings)
+    df["label"] = labels
+    df.to_csv(f"../../data/gold/{filename}.csv", index=False)
+
+
+if __name__ == "__main__":
+    generate_embeddings("train_random_embeddings", 300, 900)
+    generate_embeddings("valid_random_embeddings", 300, 100)
+    generate_embeddings("test_random_embeddings",  300, 1000)
